requirements.txt
<<<<<<< HEAD
streamlit==1.28.0
pandas==2.1.0
requests==2.31.0
google-generativeai==0.3.1
plotly==5.17.0
openpyxl==3.1.2
=======
streamlit==1.28.0
pandas==2.1.0
requests==2.31.0
google-generativeai==0.3.1
plotly==5.17.0
openpyxl==3.1.2
>>>>>>> 682e4b9c (Initial commit of Project Samarth)
python-dotenv==1.0.0